Move message to api/chat_ops controller
